chore: mv tls.go into config.go
